webapp/backend/requirements.txt
google-generativeai
fastapi
uvicorn
Pillow
python-multipart
